refactor dps results to multiple components
