+++ frontend/osrs-damage-sim/src/app/features/dps-calc/dps-results/dps-graph-tabs/dps-graph-tabs.component.html
@@ -1,0 +1,40 @@
+<div>
+  <nav class="navbar navbar-expand-sm bg-light mb-3 rounded">
+    <div class="container-fluid px-2">
+      <ul class="navbar-nav nav-tabs flex-wrap">
+        <ng-container *ngFor="let resultTab of resultTabs">
+          <li class="nav-item">
+            <a
+              class="nav-link"
+              role="button"
+              [ngClass]="{
+                active: activeResultTab.resultType === resultTab.resultType
+              }"
+              (click)="activeResultTab = resultTab"
+            >
+              <div class="text-center">{{ resultTab.label }}</div>
+            </a>
+          </li>
+        </ng-container>
+      </ul>
+    </div>
+  </nav>
+
+  <div *ngIf="activeResultTab.resultType === ResultType.DpsGraph">
+    <app-dps-graph [dpsGrapherResults]="dpsGrapherResults" [inputSetup]="inputSetup"></app-dps-graph>
+  </div>
+  <div *ngIf="activeResultTab.resultType === ResultType.HitDist">
+    <app-hit-dist-graph
+      [dpsCalcResults]="dpsCalcResults"
+      [inputSetup]="inputSetup"
+      [showResultTextLabel]="showResultTextLabel$ | async"
+    ></app-hit-dist-graph>
+  </div>
+  <div *ngIf="activeResultTab.resultType === ResultType.CalcDetails">
+    <app-calc-details
+      [dpsCalcResults]="dpsCalcResults"
+      [inputSetup]="inputSetup"
+      [showResultTextLabel]="showResultTextLabel$ | async"
+    ></app-calc-details>
+  </div>
+</div>
