+++ frontend/osrs-damage-sim/src/app/features/dps-calc/dps-results/dps-graph-tabs/dps-graph-tabs.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { DpsGraphTabsComponent } from './dps-graph-tabs.component';
+
+describe('DpsGraphTabsComponent', () => {
+  let component: DpsGraphTabsComponent;
+  let fixture: ComponentFixture<DpsGraphTabsComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [DpsGraphTabsComponent]
+    })
+    .compileComponents();
+    
+    fixture = TestBed.createComponent(DpsGraphTabsComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
