+++ frontend/osrs-damage-sim/src/app/features/dps-calc/dps-results/dps-results.component.html
@@ -1,97 +1,15 @@
 <div *ngIf="dpsResults?.dpsCalcResults && !dpsResults?.error">
-  <h3 class="mb-3 fw-light text-center">{{ dpsResults.dpsCalcResults.title }}</h3>
+  <h3 class="mb-3 fw-light text-center">{{ dpsResults.dpsCalcResults[0].title }}</h3>
 
-  <div class="row mb-1">
-    <table class="table table-bordered border-dark mb-3">
-      <thead>
-        <tr class="text-end">
-          <th scope="col" style="width: 30%" class="text-start">
-            <div class="d-flex">
-              <div class="flex-fill">Setup Name</div>
-            </div>
-          </th>
-          <ng-container *ngFor="let dpsSortField of dpsSortFields">
-            <th scope="col" style="width: 10%" (click)="sortDpsResults(dpsSortField)">
-              <button type="button" class="btn w-100 fw-bold p-0 text-end">
-                {{ sortLabels[dpsSortField] }}
-                <ng-container *ngIf="sortConfigs[dpsSortField].isSorted; else notSorted">
-                  <ng-container *ngIf="sortConfigs[dpsSortField].sortOrder === SortOrder.Ascending">
-                    <i class="bi bi-sort-up text-primary"></i>
-                  </ng-container>
-                  <ng-container *ngIf="sortConfigs[dpsSortField].sortOrder === SortOrder.Descending">
-                    <i class="bi bi-sort-down-alt text-primary"></i>
-                  </ng-container>
-                </ng-container>
-                <ng-template #notSorted>
-                  <i class="bi bi-filter text-secondary"></i>
-                </ng-template>
-              </button>
-            </th>
-          </ng-container>
-        </tr>
-      </thead>
-      <tbody>
-        <ng-container *ngFor="let i of sortIndexOrder">
-          <ng-container *ngIf="dpsResults.dpsCalcResults.results[i]; let result">
-            <tr class="align-middle text-end">
-              <th scope="row" class="text-start">
-                <app-input-setup-label
-                  [showTextLabel]="showResultTextLabel$ | async"
-                  [dpsCalcResult]="result"
-                  [inputGearSetup]="inputSetup.inputGearSetups[i]"
-                ></app-input-setup-label>
-              </th>
-              <td>{{ result.theoreticalDps | number : "1.3-3" }}</td>
-              <td>{{ result.ticksToKill | tickToTime }}</td>
-              <td>{{ result.maxHit | maxHit }}</td>
-              <td>{{ result.accuracy | number : "1.2-2" }}%</td>
-              <td>{{ result.expectedHit | number : "1.1-1" }}</td>
-              <td>{{ result.attackRoll }}</td>
-            </tr>
-          </ng-container>
-        </ng-container>
-      </tbody>
-    </table>
+  <div class="mb-1">
+    <app-dps-table [dpsCalcResults]="dpsResults.dpsCalcResults[0]" [inputSetup]="inputSetup"></app-dps-table>
   </div>
 
-  <div class="mb-3">
-    <nav class="navbar navbar-expand-sm bg-light mb-3 rounded">
-      <div class="container-fluid px-2">
-        <ul class="navbar-nav nav-tabs flex-wrap">
-          <ng-container *ngFor="let resultTab of resultTabs">
-            <li class="nav-item">
-              <a
-                class="nav-link"
-                role="button"
-                [ngClass]="{
-                  active: activeResultTab.resultType === resultTab.resultType
-                }"
-                (click)="activeResultTab = resultTab"
-              >
-                <div class="text-center">{{ resultTab.label }}</div>
-              </a>
-            </li>
-          </ng-container>
-        </ul>
-      </div>
-    </nav>
-
-    <div *ngIf="activeResultTab.resultType === ResultType.DpsGraph">
-      <app-dps-graph [dpsGrapherResults]="dpsResults.dpsGrapherResults" [inputSetup]="inputSetup"></app-dps-graph>
-    </div>
-    <div *ngIf="activeResultTab.resultType === ResultType.HitDist">
-      <app-hit-dist-graph
-        [dpsCalcResults]="dpsResults.dpsCalcResults"
-        [inputSetup]="inputSetup"
-        [showResultTextLabel]="showResultTextLabel$ | async"
-      ></app-hit-dist-graph>
-    </div>
-    <div *ngIf="activeResultTab.resultType === ResultType.CalcDetails">
-      <app-calc-details
-        [dpsCalcResults]="dpsResults.dpsCalcResults"
-        [inputSetup]="inputSetup"
-        [showResultTextLabel]="showResultTextLabel$ | async"
-      ></app-calc-details>
-    </div>
+  <div class="mb-1">
+    <app-dps-graph-tabs
+      [dpsCalcResults]="dpsResults.dpsCalcResults[0]"
+      [dpsGrapherResults]="dpsResults.dpsGrapherResults[0]"
+      [inputSetup]="inputSetup"
+    ></app-dps-graph-tabs>
   </div>
 </div>
