+++ frontend/osrs-damage-sim/src/app/shared/components/npc-input/npc-input.component.html
@@ -7,6 +7,7 @@
         [allValues]="allNpcs"
         [selectedValue]="npcInfo.npc"
         [valueFilter]="npcFilter"
+        [clearable]="false"
         placeholder="Select npc"
         (valueChanged)="selectedNpcChange($event)"
       >
